pyproject.toml
[tool.poetry]
name = "plannotate"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
biopython = "^1.84"
click = {version = "8.0.4", extras = ["cli", "webapp"]}
protobuf = {version = "3.20.3", extras = ["webapp"]}
streamlit = {version = "1.8.1", extras = ["webapp"]}
altair = {version = "4.2.2", extras = ["webapp"]}
bokeh = {version = "^3.6.1", extras = ["webapp"]}
fastapi = {version = "^0.115.5", extras = ["api"]}
uvicorn = {version = "^0.32.1", extras = ["api"]}
python-multipart = {version = "^0.0.17", extras = ["api"]}


[tool.poetry.group.dev.dependencies]
watchfiles = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
